requirements.txt
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
bcrypt==4.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
dnspython==2.2.1
email-validator==1.3.0
executing==1.1.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.1
Flask-WTF==1.0.1
idna==3.4
ipdb==0.13.9
ipython==8.5.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
stack-data==0.5.1
toml==0.10.2
traitlets==5.4.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
